Adding supervisors

+++ app/datatables/effective_work_experience_summaries_datatable.rb
@@ -5,12 +5,14 @@ class EffectiveWorkExperienceSummariesDatatable < Effective::Datatable
 
     col :token, visible: false
     col :created_at, visible: false
+    col :user, visible: false
 
     col :start_on
     col :end_on, visible: false
     col :period
 
     col :mentor
+    col :supervisor
 
     col(:total_hours, label: 'Hours') do |work_experience_summary|
       work_experience_hours_to_s(work_experience_summary.total_hours)
@@ -31,9 +33,11 @@ class EffectiveWorkExperienceSummariesDatatable < Effective::Datatable
     end
 
     actions_col(show: false) do |work_experience_summary|
-      if work_experience_summary.draft?
+      if work_experience_summary.draft? && EffectiveResources.authorized?(self, :update, work_experience_summary)
         dropdown_link_to('Continue', effective_work_experience.work_experience_summary_build_path(work_experience_summary, work_experience_summary.next_step), 'data-turbolinks' => false, 'data-turbo' => false)
-        dropdown_link_to('Delete', effective_work_experience.work_experience_summary_path(work_experience_summary), 'data-confirm': "Really delete #{work_experience_summary}?", 'data-method': :delete)
+        if EffectiveResources.authorized?(self, :destroy, work_experience_summary)
+          dropdown_link_to('Delete', effective_work_experience.work_experience_summary_path(work_experience_summary), 'data-confirm': "Really delete #{work_experience_summary}?", 'data-method': :delete)
+        end
       else
         dropdown_link_to('Show', effective_work_experience.work_experience_summary_path(work_experience_summary))
       end
@@ -41,9 +45,7 @@ class EffectiveWorkExperienceSummariesDatatable < Effective::Datatable
   end
 
   collection do
-    scope = EffectiveWorkExperience.WorkExperienceSummary.deep.all
-    scope = scope.where(user_id: attributes[:user_id], user_type: attributes[:user_type]) if attributes[:user_id].present?
-    scope
+    EffectiveWorkExperience.WorkExperienceSummary.deep.all
   end
 
 end
